Added GIMP scripts for working on Sierra-style images
Sierra AGI had a resolution such that pixels were not square - they had an 8:5 aspect ratio.
Now that pixels are generally square, these pixels are approximated with a 2:1 aspect ratio.
These scripts make it easy to downsample an image so that all the pixels become rectangles,
or to save an image - including all its layers - as a fanned out single layer PNG - and
with half the horizontal resolution, since every other pixel is redundant.